refactor(constants): Standardize POWERUP_DURATION data type to int

+++ Arkanoid/main/util/Constants.java
@@ -26,5 +26,5 @@ public final class Constants {
 
     //Bổ sung thêm 2 hằng số: Độ dài tăng thêm thanh + Thời gian 7s hiệu lực PowerUp.
     public static final int PADDLE_EXPAND_AMOUNT = 20;
-    public static final long POWERUP_DURATION = 7000; //7000 milis.
+    public static final int POWERUP_DURATION = 7000; //7000 milis.
 }
